refactor: DTO 내 Long 타입을 Integer로 일괄 수정

+++ src/main/java/org/scoula/domain/chatbot/dto/ChatResponseDto.java
@@ -14,6 +14,6 @@ import lombok.NoArgsConstructor;
 public class ChatResponseDto {
     private String content; //gpt 응답
     private String intentType; // 응답 intent (백엔드 판단 결과일 수도 있음 -> 추후 확장?)
-    private Long messageId; // 저장된 chat_messages.id (있으면)
+    private Integer messageId; // 저장된 chat_messages.id (있으면)
 
 }
